* pan/otp_8332_halfword:
  Teach testcase in driver_suite the new prototype for driver_async
  wx: Correct usage of driver callbacks from wx thread
  Adopt the new (R13B04) Nif functionality to the halfword codebase
  Support monitoring and demonitoring from driver threads
  Fix further test-suite problems
  Correct the VM to work for more test suites
  Teach {wordsize,internal|external} to system_info/1
  Make tracing and distribution work
  Turn on instruction packing in the loader and virtual machine
  Add the BeamInstr data type for loaded BEAM code
  Fix the BEAM dissambler for the half-word emulator
  Store pointers to heap data in 32-bit words
  Add a custom mmap wrapper to force heaps into the lower address range
  Fit all heap data into the 32-bit address range

+++ erts/emulator/beam/erl_async.c
@@ -1,19 +1,19 @@
 /*
  * %CopyrightBegin%
- * 
- * Copyright Ericsson AB 2000-2009. All Rights Reserved.
- * 
+ *
+ * Copyright Ericsson AB 2000-2010. All Rights Reserved.
+ *
  * The contents of this file are subject to the Erlang Public License,
  * Version 1.1, (the "License"); you may not use this file except in
  * compliance with the License. You should have received a copy of the
  * Erlang Public License along with this software. If not, it can be
  * retrieved online at http://www.erlang.org/.
- * 
+ *
  * Software distributed under the License is distributed on an "AS IS"
  * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See
  * the License for the specific language governing rights and limitations
  * under the License.
- * 
+ *
  * %CopyrightEnd%
  */
 #ifdef HAVE_CONFIG_H
@@ -390,14 +390,14 @@ int check_async_ready(void)
 **      async_data     data to pass to invoke function
 **      async_free     function for relase async_data in case of failure
 */
-long driver_async(ErlDrvPort ix, unsigned int* key,
+long driver_async(ErlDrvPort ix, unsigned long* key,
 		  void (*async_invoke)(void*), void* async_data,
 		  void (*async_free)(void*))
 {
     ErlAsync* a = (ErlAsync*) erts_alloc(ERTS_ALC_T_ASYNC, sizeof(ErlAsync));
     Port* prt = erts_drvport2port(ix);
     long id;
-    unsigned int qix;
+    unsigned long qix;
 
 
     if (!prt)
